+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Additions
 Fixes
 - Fixed tooltips not being bounded correctly for viewports that don't start at the origin
 - Fixed a stack overflow exception when a panel's children have just enough height to cause a scroll bar to appear
+- Fixed AddedToUi and RemovedFromUi being called for freshly added or removed children whose parents are not in a ui system
 
 ### MLEM.Data
 Improvements